backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
boto3==1.28.27
python-dotenv==1.0.0
requests==2.32.3
cachetools==5.5.0
